+++ extensions/git/src/api/api1.ts
@@ -15,6 +15,7 @@ import { GitExtensionImpl } from './extension';
 import { GitBaseApi } from '../git-base';
 import { PickRemoteSourceOptions } from '../typings/git-base';
 import { OperationKind, OperationResult } from '../operation';
+import { CloneUtil } from '../clone';
 
 class ApiInputBox implements InputBox {
 	#inputBox: SourceControlInputBox;
@@ -401,7 +402,7 @@ export class ApiImpl implements API {
 
 	async clone(uri: Uri, options?: CloneOptions): Promise<Uri | null> {
 		const parentPath = options?.parentPath?.fsPath;
-		const result = await this.#model.clone(uri.toString(), { parentPath, recursive: options?.recursive, ref: options?.ref, postCloneAction: options?.postCloneAction, skipCache: options?.skipCache });
+		const result = await CloneUtil.clone(this.#model.git, this.#model.telemetryReporter, this.#model.repositoryCache, uri.toString(), { parentPath, recursive: options?.recursive, ref: options?.ref, postCloneAction: options?.postCloneAction, skipCache: options?.skipCache });
 		return result ? Uri.file(result) : null;
 	}
